pdf_parser/pyproject.toml
[project]
name = "pdf-parser"
version = "0.1.0"
description = "Universal PDF to Markdown parser with table extraction"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "click>=8.1.7",
    "pillow>=10.0.0",
    "anthropic>=0.40.0",
]

[project.scripts]
pdf-parser = "pdf_parser.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]
